# Conflicts:
#	aplanmis/src/main/java/com/augurit/efficiency/service/impl/WinEfficiencySupervisionServiceImpl.java

+++ aplanmis-mall/src/main/java/com/augurit/aplanmis/mall/userCenter/controller/RestUserCenterController.java
@@ -6,13 +6,16 @@ import com.augurit.agcloud.framework.util.StringUtils;
 import com.augurit.aplanmis.common.domain.AeaLinkmanInfo;
 import com.augurit.aplanmis.common.domain.AeaUnitInfo;
 import com.augurit.aplanmis.common.domain.AeaUnitLinkman;
+import com.augurit.aplanmis.common.mapper.AeaLinkmanInfoMapper;
 import com.augurit.aplanmis.common.mapper.AeaUnitLinkmanMapper;
 import com.augurit.aplanmis.common.service.linkman.AeaLinkmanInfoService;
 import com.augurit.aplanmis.common.service.unit.AeaUnitInfoService;
 import com.augurit.aplanmis.common.utils.SessionUtil;
-import com.augurit.aplanmis.common.vo.AeaUnitInfoVo;
 import com.augurit.aplanmis.common.vo.LoginInfoVo;
 import com.augurit.aplanmis.mall.userCenter.constant.LoginUserRoleEnum;
+import com.augurit.aplanmis.mall.userCenter.service.RestUserCenterService;
+import com.augurit.aplanmis.mall.userCenter.vo.AeaLinkmanInfoVo;
+import com.augurit.aplanmis.mall.userCenter.vo.AeaUnitInfoVo;
 import com.augurit.aplanmis.mall.userCenter.vo.UserInfoVo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
@@ -45,6 +48,10 @@ public class RestUserCenterController {
     private AeaLinkmanInfoService aeaLinkmanInfoService;
     @Autowired
     private AeaUnitLinkmanMapper aeaUnitLinkmanMapper;
+    @Autowired
+    private RestUserCenterService restUserCenterService;
+    @Autowired
+    private AeaLinkmanInfoMapper aeaLinkmanInfoMapper;
 
     @GetMapping("/toMyCloundSpacesPage")
     @ApiOperation(value = "法人空间我的首页")
@@ -114,15 +121,20 @@ public class RestUserCenterController {
     @ApiOperation("保存单位联系人信息及单位与联系人的关系")
     public ResultForm saveLinkmanAndUnitLinkman(AeaLinkmanInfo aeaLinkmanInfo, HttpServletRequest request){
         //要求：登录名，密码字段不用填，unitInfoId不为空
+        LoginInfoVo user = SessionUtil.getLoginInfo(request);
         if(StringUtils.isEmpty(aeaLinkmanInfo.getUnitInfoId()))  return new ResultForm(false,"单位为空");
+        //判断传参单位是否和后台缓存单位一致
+        if (!aeaLinkmanInfo.getUnitInfoId().equals(user.getUnitId())) return new ResultForm(false,"保存单位联系人信息及单位与联系人的关系失败");
         try {
             if(StringUtils.isNotBlank(aeaLinkmanInfo.getLinkmanInfoId())){
+                if (!restUserCenterService.isBelongUnit(aeaLinkmanInfo.getLinkmanInfoId(),request)) return new ResultForm(false,"保存单位联系人信息及单位与联系人的关系失败");
                 aeaLinkmanInfoService.updateAeaLinkmanInfo(aeaLinkmanInfo);
             }else{//根据身份证号码查询联系人，若已存在，则更新信息，若不存在，则新增
                 AeaLinkmanInfo query=new AeaLinkmanInfo();
                 query.setLinkmanCertNo(aeaLinkmanInfo.getLinkmanCertNo());
                 List<AeaLinkmanInfo> linkmans = aeaLinkmanInfoService.findLinkmanInfo(query);
                 if(linkmans.size()>0){
+                    if (!restUserCenterService.isBelongUnit(aeaLinkmanInfo.getLinkmanInfoId(),request)) return new ResultForm(false,"保存单位联系人信息及单位与联系人的关系失败");
                     aeaLinkmanInfo.setLinkmanInfoId(linkmans.get(0).getLinkmanInfoId());
                     aeaLinkmanInfoService.updateAeaLinkmanInfo(aeaLinkmanInfo);
                 }else{
@@ -144,7 +156,7 @@ public class RestUserCenterController {
             }
         } catch (Exception e) {
             logger.error(e.getMessage(),e);
-            return new ResultForm(false,"保存联系人信息失败");
+            return new ResultForm(false,"发生错误");
         }
         return new ResultForm(true,"success");
     }
@@ -155,17 +167,19 @@ public class RestUserCenterController {
     @ApiImplicitParams({@ApiImplicitParam(value = "单位Id",name = "unitInfoId",required = true,dataType = "string"),
             @ApiImplicitParam(value = "联系人Id",name = "linkmanInfoId",required = true,dataType = "string"),
             @ApiImplicitParam(value = "绑定关系0:解绑，1:绑定",name = "isBindAccount",required = true,dataType = "string")})
-    public ResultForm bindOrUnbindRelation(String unitInfoId,String linkmanInfoId,String isBindAccount){
+    public ResultForm bindOrUnbindRelation(String unitInfoId,String linkmanInfoId,String isBindAccount,HttpServletRequest request){
         AeaUnitLinkman aeaUnitLinkman = new AeaUnitLinkman();
         aeaUnitLinkman.setLinkmanInfoId(linkmanInfoId);
         aeaUnitLinkman.setUnitInfoId(unitInfoId);
         //isBindAccount=0解绑  isBindAccount=1绑定
         aeaUnitLinkman.setIsBindAccount(isBindAccount);
         try {
+            //权限校验
+            if (!restUserCenterService.isBelongUnit(linkmanInfoId,request)) return new ResultForm(false,"发生错误");
             aeaLinkmanInfoService.updateAeaUnitLinkmanByUnitAndLinkman(aeaUnitLinkman);
         } catch (Exception e) {
             logger.error(e.getMessage(),e);
-            return new ResultForm(false,e.getMessage());
+            return new ResultForm(false,"发生错误");
         }
         return new ResultForm(true,"success");
     }
@@ -175,8 +189,15 @@ public class RestUserCenterController {
     @ApiOperation(value = "删除联系人")
     @ApiImplicitParams({@ApiImplicitParam(value = "单位id",name = "unitInfoId",required = true,dataType = "string"),
             @ApiImplicitParam(value = "联系人id",name = "linkmanInfoId",required = true,dataType = "string")})
-    public ResultForm deleteRelation(@PathVariable("unitInfoId") String unitInfoId,@PathVariable("linkmanInfoId") String linkmanInfoId){
-        aeaLinkmanInfoService.deleteUnitLinkman(unitInfoId,linkmanInfoId);
+    public ResultForm deleteRelation(@PathVariable("unitInfoId") String unitInfoId,@PathVariable("linkmanInfoId") String linkmanInfoId,HttpServletRequest request){
+        try {
+            //权限校验
+            if (!restUserCenterService.isBelongUnit(linkmanInfoId,request)) return new ResultForm(false,"发生错误");
+            aeaLinkmanInfoService.deleteUnitLinkman(unitInfoId,linkmanInfoId);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResultForm(false,"发生错误");
+        }
         return new ResultForm(true,"success");
     }
 
@@ -186,11 +207,11 @@ public class RestUserCenterController {
     @ApiImplicitParam(value = "空值:查询所有人，1:查询委托人",name = "isAll",required = true,dataType = "string")
     public ResultForm getAllUnitLinkInfo(HttpServletRequest request, String isAll) throws Exception{
         LoginInfoVo user = SessionUtil.getLoginInfo(request);
-        List<AeaLinkmanInfo> linkmanInfoList = new ArrayList<>();
+        List<AeaLinkmanInfo> aeaLinkmanInfos = new ArrayList<>();
         if(user != null){
-            linkmanInfoList=aeaLinkmanInfoService.getAeaLinkmanInfoByUnitInfoIdAndIsBindAccount(user.getUnitId(),isAll);
+            aeaLinkmanInfos =  aeaLinkmanInfoMapper.getAeaLinkmanInfoByUnitInfoIdAndIsBindAccount(user.getUnitId(),isAll);
         }
-        return new ContentResultForm<>(true,linkmanInfoList);
+        return new ContentResultForm<>(true,aeaLinkmanInfos.stream().map(AeaLinkmanInfoVo::build).collect(Collectors.toList())) ;
     }
 
     //查询某单位的所有联系人列表
@@ -204,53 +225,62 @@ public class RestUserCenterController {
 
     @GetMapping("userinfo")
     @ApiOperation(value = "获取当前登录用户详情")
-    public ResultForm getUserInfo(HttpServletRequest request){
+    public ContentResultForm<UserInfoVo> getUserInfo(HttpServletRequest request){
         UserInfoVo userInfoVo = new UserInfoVo();
         LoginInfoVo user = SessionUtil.getLoginInfo(request);
-        if(user!=null){
-            if("1".equals(user.getIsPersonAccount())){//个人
-                AeaLinkmanInfo aeaLinkmanInfo=aeaLinkmanInfoService.getAeaLinkmanInfoByLinkmanInfoId(user.getUserId());
-                userInfoVo.setAeaLinkmanInfo(aeaLinkmanInfo);
-                userInfoVo.setRole(LoginUserRoleEnum.PERSONAL.getValue());
-            }else if(StringUtils.isNotBlank(user.getUserId())){//委托人
-                AeaLinkmanInfo aeaLinkmanInfo=aeaLinkmanInfoService.getAeaLinkmanInfoByLinkmanInfoId(user.getUserId());
-                List<AeaUnitInfo> aeaUnitList=aeaUnitInfoService.getUnitInfoByLinkmanInfoId(user.getUserId());
-                List<AeaUnitInfoVo> vos = aeaUnitList.stream().map(AeaUnitInfoVo::build).collect(Collectors.toList());
-                userInfoVo.setAeaUnitList(vos);
-                userInfoVo.setAeaLinkmanInfo(aeaLinkmanInfo);
-                userInfoVo.setRole(LoginUserRoleEnum.HANDLE.getValue());
-            }else{//企业
-                AeaUnitInfo aeaUnitInfo = aeaUnitInfoService.getAeaUnitInfoByUnitInfoId(user.getUnitId());
-                List<AeaLinkmanInfo> linkmanInfoList=aeaLinkmanInfoService.findAllUnitLinkman(user.getUnitId());
-                userInfoVo.setLinkmanInfoList(linkmanInfoList);
-                userInfoVo.setAeaUnitInfo(AeaUnitInfoVo.build(aeaUnitInfo));
-                userInfoVo.setRole(LoginUserRoleEnum.UNIT.getValue());
+
+        try {
+            if(user!=null){
+                if("1".equals(user.getIsPersonAccount())){//个人
+                    AeaLinkmanInfoVo aeaLinkmanInfo=restUserCenterService.getAeaLinkmanInfoByLinkmanInfoId(user.getUserId());
+                    userInfoVo.setAeaLinkmanInfo(aeaLinkmanInfo);
+                    userInfoVo.setRole(LoginUserRoleEnum.PERSONAL.getValue());
+                }else if(StringUtils.isNotBlank(user.getUserId())){//委托人
+                    AeaLinkmanInfoVo aeaLinkmanInfo=restUserCenterService.getAeaLinkmanInfoByLinkmanInfoId(user.getUserId());
+                    List<AeaUnitInfoVo> aeaUnitList=restUserCenterService.getUnitInfoListByLinkmanInfoId(user.getUserId());
+                    userInfoVo.setAeaUnitList(aeaUnitList);
+                    userInfoVo.setAeaLinkmanInfo(aeaLinkmanInfo);
+                    userInfoVo.setRole(LoginUserRoleEnum.HANDLE.getValue());
+                }else{//企业
+                    AeaUnitInfoVo aeaUnitInfo = restUserCenterService.getAeaUnitInfoByUnitInfoId(user.getUnitId());
+                    List<AeaLinkmanInfoVo> linkmanInfoList=restUserCenterService.findAllUnitLinkman(user.getUnitId());
+                    userInfoVo.setLinkmanInfoList(linkmanInfoList);
+                    userInfoVo.setAeaUnitInfo(aeaUnitInfo);
+                    userInfoVo.setRole(LoginUserRoleEnum.UNIT.getValue());
+                }
             }
+            return new ContentResultForm<>(true,userInfoVo);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ContentResultForm<>(false,null,"获取当前登录用户详情发生异常");
         }
-        return new ContentResultForm<>(true,userInfoVo);
     }
 
     @GetMapping("unitinfo/list/{linkmanInfoId}")
     @ApiOperation(value = "查询联系人单位列表")
     @ApiImplicitParam(value = "联系人id",name = "linkmanInfoId",required = true,dataType = "string")
     public ResultForm getUnitInfoByLinkmanInfoId (@PathVariable("linkmanInfoId") String linkmanInfoId) throws Exception {
-        return new ContentResultForm<>(true,aeaUnitInfoService.getUnitInfoByLinkmanInfoId(linkmanInfoId)) ;
+        return new ContentResultForm<>(true,restUserCenterService.getUnitInfoListByLinkmanInfoId(linkmanInfoId)) ;
     }
 
     @GetMapping("linkman/{linkmanInfoId}/{unitInfoId}")
     @ApiOperation(value = "查询联系人信息")
     @ApiImplicitParams({@ApiImplicitParam(value = "单位id",name = "unitInfoId",required = true,dataType = "string"),
                 @ApiImplicitParam(value = "联系人id",name = "linkmanInfoId",required = true,dataType = "string")})
-    public ResultForm getByLinkmanInfoId (@PathVariable("linkmanInfoId") String linkmanInfoId,@PathVariable("unitInfoId") String unitInfoId) {
+    public ResultForm getByLinkmanInfoId (@PathVariable("linkmanInfoId") String linkmanInfoId,@PathVariable("unitInfoId") String unitInfoId,HttpServletRequest request) {
         if(StringUtils.isBlank(linkmanInfoId)) return new ResultForm(false,"联系人参数缺失");
         if(StringUtils.isBlank(unitInfoId)) return new ResultForm(false,"单位参数缺失");
-        AeaLinkmanInfo aeaLinkmanInfo = aeaLinkmanInfoService.getAeaLinkmanInfoByLinkmanInfoId(linkmanInfoId);
+
+        List<AeaUnitLinkman> unitLinkmans = new ArrayList<>();
+        AeaLinkmanInfoVo aeaLinkmanInfo = new AeaLinkmanInfoVo();
+        try {
+            //权限校验
+        if (!restUserCenterService.isBelongUnit(linkmanInfoId,request)) return new ResultForm(false,"查找单位下的联系人信息失败");
+        aeaLinkmanInfo=restUserCenterService.getAeaLinkmanInfoByLinkmanInfoId(linkmanInfoId);
         AeaUnitLinkman query=new AeaUnitLinkman();
         query.setUnitInfoId(unitInfoId);
         query.setLinkmanInfoId(linkmanInfoId);
-        List<AeaUnitLinkman> unitLinkmans = new ArrayList<>();
-        try {
-            unitLinkmans = aeaUnitLinkmanMapper.listAeaUnitLinkman(query);
+        unitLinkmans = aeaUnitLinkmanMapper.listAeaUnitLinkman(query);
         } catch (Exception e) {
             logger.error(e.getMessage(),e);
             return new ResultForm(false,"查找单位下的联系人信息失败");
@@ -259,16 +289,20 @@ public class RestUserCenterController {
         return new ContentResultForm<>(true,aeaLinkmanInfo) ;
     }
 
-    @GetMapping("linkman/list/key")
-    @ApiOperation(value = "根据姓名，电话，身份证号等关键字查询联系人列表")
-    @ApiImplicitParams({@ApiImplicitParam(value = "姓名，电话，身份证号等关键字",name = "keyword",required = false,dataType = "string"),
-            @ApiImplicitParam(value = "单位信息ID",name = "unitInfoId",required = false,dataType = "string")})
-    public ResultForm getByIdCard (String keyword, String unitInfoId) throws Exception {
-        return new ContentResultForm<>(true,aeaLinkmanInfoService.getByKeyword(keyword,unitInfoId)) ;
-    }
+
 
     /*******************************************企业中心模块数据请求结束*****************************************************/
 
 
+
+
+//    @GetMapping("linkman/list/key")
+//    @ApiOperation(value = "根据姓名，电话，身份证号等关键字查询联系人列表")
+//    @ApiImplicitParams({@ApiImplicitParam(value = "姓名，电话，身份证号等关键字",name = "keyword",required = false,dataType = "string"),
+//            @ApiImplicitParam(value = "单位信息ID",name = "unitInfoId",required = false,dataType = "string")})
+//    public ResultForm getByIdCard (String keyword, String unitInfoId) throws Exception {
+//        List<AeaLinkmanInfo> aeaLinkmanInfos = aeaLinkmanInfoService.getByKeyword(keyword,unitInfoId);
+//        return new ContentResultForm<>(true,aeaLinkmanInfos.stream().map(AeaLinkmanInfoVo::build).collect(Collectors.toList())) ;
+//    }
 
 }
